+++ src/Symfony/Bridge/PhpUnit/CHANGELOG.md
@@ -7,6 +7,7 @@ CHANGELOG
  * ignore verbosity settings when the build fails because of deprecations
  * added per-group verbosity
  * added `ExpectDeprecationTrait` to be able to define an expected deprecation from inside a test
+ * deprecated the `@expectedDeprecation` annotation, use the `ExpectDeprecationTrait::expectDeprecation()` method instead
 
 5.0.0
 -----